Add home_dir_perms, create_subdirs and subdir_perms
The `home_dir_perms`new variable allows to create the sftpgo user homedir with
special permissions. It is specially useful when you don't want to allow the
user to create directories in base homedir.

The `create_subdirs` and `subdir_perms` vars allow to create subdirs in sftpgo
user homedir and define the permissions. It is useful when setting
home_dir_perms with something like 0555, this way the user cannot create dirs
or files in homedir directly. Creating subdirs with something like '0550'
doesn't allow to delete or rename the subdirs.